Use reflection to call the method in SM. (#3188)
* Use reflection to call the method in SM.

* fix error on linux

* Add log for debug

* fix typo

* Use System.ServiceModel.Primitives.dll instead of system.servicemodel.primitives.dll

* Fix casesensitive.

* Add some check for the code that use reflection.

* fix the path error.

* Code change based on Matt's feedback.

* fix errors.

* Move Find method out of the loop.

* Add some comments. Cast operations as IEnumberable.

* Some rename change.

+++ src/svcutilcore/tests/SvcutilTests.cs
@@ -20,9 +20,14 @@ namespace Microsoft.Tools.ServiceModel.SvcUtil.XmlSerializer.Tests
         {
             string outputFile = Path.GetTempFileName() + ".cs";
             Assert.False(File.Exists(outputFile));
-            Tool.Main(new string[] { Assembly.GetExecutingAssembly().Location, $"/out:{outputFile}" });
-            Assert.True(File.Exists(outputFile));
-            File.Delete(outputFile);
+            string testassemblypath = Assembly.GetExecutingAssembly().Location;
+            string smassemblypath = testassemblypath.Replace("dotnet-svcutil.xmlserializer.Tests", "System.ServiceModel.Primitives").Replace("Microsoft.Tools.ServiceModel.SvcUtil.XmlSerializer.Tests.dll", "System.ServiceModel.Primitives.dll").Replace("netcoreapp", "netstandard");
+            if (File.Exists(smassemblypath))
+            {
+                Tool.Main(new string[] { Assembly.GetExecutingAssembly().Location, $"/out:{outputFile}", $"/smreference:{smassemblypath}" });
+                Assert.True(File.Exists(outputFile));
+                File.Delete(outputFile);
+            }
         }
     }
 }
